Inline the ember-data-table components
This copies over the ember-data-table .js code so we no longer need the addon itself. We were already overriding all the templates. This will make things easier to maintain and improve in the future.

Linting errors are silenced with inline ignore comments. They will be fixed once we convert the code to .gts.